requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.2
flask-migrate==4.0.5
flask-wtf==1.1.1
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
Werkzeug==2.3.7
Pillow==10.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-weather==0.1.0
SQLAlchemy==2.0.20 